requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1

# Document processing
pypdf==3.17.0
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1
chardet==5.2.0

# Embeddings and ML
openai==1.3.7
sentence-transformers==2.2.2
torch==2.1.1
numpy==1.24.3

# Vector database (local only for POC)
chromadb==0.4.18


# Utilities
python-dotenv==1.0.0
httpx==0.25.0
tenacity==8.2.3
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
ruff==0.1.6
mypy==1.7.1
black==23.11.0
flake8==6.1.0
ipython==8.18.0